% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{input-methods}
\alias{input}
\alias{input,MLSeq-method}
\alias{input-methods}
\title{Accessors for the 'inputObject' slot of an \code{MLSeq} object}
\usage{
input(object)

\S4method{input}{MLSeq}(object)
}
\arguments{
\item{object}{an \code{MLSeq} object.}
}
\description{
\code{MLSeq} package benefits from \code{DESeqDataSet} structure from bioconductor package \code{DESeq2} for storing gene
expression data in a comprehensive structure. This object is used as an input for classification task through \code{\link{classify}}.
The input is stored in \code{inputObject} slot of \code{MLSeq} object.
}
\examples{
library(DESeq2)
data(cervical)

# a subset of cervical data with first 150 features.
data <- cervical[c(1:150),]

# defining sample classes.
class <- data.frame(condition=factor(rep(c("N","T"), c(29,29))))

n <- ncol(data)  # number of samples
p <- nrow(data)  # number of features

# number of samples for test set (20\% test, 80\% train).
nTest <- ceiling(n*0.2)
ind <- sample(n, nTest, FALSE)

# train set
data.train <- data[,-ind]
data.train <- as.matrix(data.train + 1)
classtr <- data.frame(condition=class[-ind, ])

# train set in S4 class
data.trainS4 <- DESeqDataSetFromMatrix(countData = data.train,
                  colData = classtr, formula(~ condition))
data.trainS4 <- DESeq(data.trainS4, fitType = "local")

# Classification and Regression Trees (CART)
cart <- classify(data = data.trainS4, method = "rpart",
          transformation = "vst", ref = "T", normalize = "deseq",
          control = trainControl(method = "repeatedcv", number = 5,
                                 repeats = 3, classProbs = TRUE))

input(cart)

}
\author{
Gokmen Zararsiz
}
\seealso{
\code{\link{classify}}, \code{\link[DESeq2]{DESeqDataSet}}
}

