% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{method-methods}
\alias{method}
\alias{method,MLSeq-method}
\alias{method-methods}
\title{Accessors for the 'method' slot of an \code{MLSeq} object}
\usage{
method(object)

\S4method{method}{MLSeq}(object)
}
\arguments{
\item{object}{an \code{MLSeq} object.}
}
\description{
This slot stores the name of selected model which is used in \code{classify} function.
The trained model is stored in slot \code{trainedModel}.
See \code{\link{trained}} for details.
}
\details{
\code{method} slot stores the name of the classification method as "svm", support vector machines using
radial-based kernel function; "bagsvm", support vector machines with bagging ensemble; "randomForest",
random forest algorithm and "cart", classification and regression trees algorithm.
}
\examples{
library(DESeq2)
data(cervical)

# a subset of cervical data with first 150 features.
data <- cervical[c(1:150),]

# defining sample classes.
class <- data.frame(condition=factor(rep(c("N","T"), c(29,29))))

n <- ncol(data)  # number of samples
p <- nrow(data)  # number of features

# number of samples for test set (20\% test, 80\% train).
nTest <- ceiling(n*0.2)
ind <- sample(n, nTest, FALSE)

# train set
data.train <- data[,-ind]
data.train <- as.matrix(data.train + 1)
classtr <- data.frame(condition=class[-ind, ])

# train set in S4 class
data.trainS4 <- DESeqDataSetFromMatrix(countData = data.train,
                  colData = classtr, formula(~ condition))
data.trainS4 <- DESeq(data.trainS4, fitType = "local")

# Classification and Regression Trees (CART)
cart <- classify(data = data.trainS4, method = "cart",
          transformation = "vst", ref = "T", normalize = "deseq",
          control = trainControl(method = "repeatedcv", number = 5,
                                 repeats = 3, classProbs = TRUE))

method(cart)

}
\author{
Gokmen Zararsiz
}
\seealso{
\code{\link{trained}}
}

