% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\docType{package}
\name{MLSeq-package}
\alias{MLSeq-package}
\title{Machine learning interface for RNA-Seq data}
\description{
This package applies several machine learning methods, including SVM, bagSVM, Random Forest and CART, to RNA-Seq data.
}
\details{
\tabular{ll}{
  Package: \tab MLSeq\cr
  Type: \tab Package\cr
  License: \tab GPL (>= 2)\cr
}
}
\author{
Gokmen Zararsiz, Dincer Goksuluk, Selcuk Korkmaz, Vahap Eldem, Izzet Parug Duru, Turgay Unver, Ahmet Ozturk

-----------------

Maintainers:

Gokmen Zararsiz, \email{gokmenzararsiz@erciyes.edu.tr}

Dincer Goksuluk \email{dincer.goksuluk@hacettepe.edu.tr}

Selcuk Korkmaz \email{selcukorkmaz@hotmail.com}
}
\keyword{package}

