% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{ref}
\alias{ref}
\title{Accessors for the 'ref' slot of an \code{MLSeq} object}
\usage{
ref(object)

\S4method{ref}{MLSeq}(object)
}
\arguments{
\item{object}{an \code{MLSeq} object.}
}
\description{
This slot stores the information about reference category. Confusion matrix and related statistics are calculated using
the user-defined reference category.
}
\examples{

data(cervical)

data <- cervical[c(1:150),]  # a subset of cervical data with first 150 features.

# defining sample classes.
class <- data.frame(condition=factor(rep(c("N", "T"), c(29, 29))))

n <- ncol(data)  # number of samples
p <- nrow(data)  # number of features

nTest <- ceiling(n*0.2)  # number of samples for test set (20\% test, 80\% train).
ind <- sample(n, nTest, FALSE)

# train set
data.train <- data[,-ind]
data.train <- as.matrix(data.train + 1)
classtr <- data.frame(condition=class[-ind, ])

# train set in S4 class
data.trainS4 <- DESeqDataSetFromMatrix(countData = data.train,
                                       colData = classtr, formula(~ condition))
data.trainS4 <- DESeq(data.trainS4, fitType = "local")

# Classification and Regression Trees (CART)
cart <- classify(data = data.trainS4, method = "cart", normalize = "deseq",
                 transformation = "vst", ref = "T",
                 control = trainControl(method = "repeatedcv", number = 5, repeats = 3, classProbs = TRUE))

ref(cart)

}
\author{
Gokmen Zararsiz
}

