% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{show-methods}
\alias{show}
\alias{show,MLSeq-method}
\alias{show-methods}
\title{Show method for MLSeq objects}
\usage{
\S4method{show}{MLSeq}(object)
}
\arguments{
\item{object}{an \code{MLSeq} object returned from \code{classify} function.}
}
\description{
Prints out the information from the trained model using \code{classify} function.
}
\author{
Gokmen Zararsiz
}
\seealso{
\code{\link{classify}}
}

