% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{normalization-methods}
\alias{normalization}
\alias{normalization,MLSeq-method}
\alias{normalization-methods}
\title{Accessors for the 'normalization' slot of an \code{MLSeq} object}
\usage{
normalization(object)

\S4method{normalization}{MLSeq}(object)
}
\arguments{
\item{object}{an \code{MLSeq} object.}
}
\description{
This slot stores the name of normalization method which is used while normalizing the count data such as "deseq", "none" or "tmm"
}
\examples{

data(cervical)

# a subset of cervical data with first 150 features.
data <- cervical[c(1:150),]

# defining sample classes.
class <- data.frame(condition=factor(rep(c("N", "T"), c(29, 29))))

n <- ncol(data)  # number of samples
p <- nrow(data)  # number of features

# number of samples for test set (20\% test, 80\% train).
nTest <- ceiling(n*0.2)
ind <- sample(n, nTest, FALSE)

# train set
data.train <- data[,-ind]
data.train <- as.matrix(data.train + 1)
classtr <- data.frame(condition=class[-ind, ])

# train set in S4 class
data.trainS4 <- DESeqDataSetFromMatrix(countData = data.train,
                                       colData = classtr, formula(~ condition))
data.trainS4 <- DESeq(data.trainS4, fitType = "local")

# Classification and Regression Trees (CART)
cart <- classify(data = data.trainS4, method = "cart", normalize = "deseq",
          transformation = "vst", ref = "T",
          control = trainControl(method = "repeatedcv", number = 5,
                                 repeats = 3, classProbs = TRUE))

normalization(cart)

}
\author{
Gokmen Zararsiz
}

